+++ tests/test_client_response.py
@@ -10,7 +10,7 @@ from yarl import URL
 
 import aiohttp
 from aiohttp import helpers
-from aiohttp.client_reqrep import ClientResponse
+from aiohttp.client_reqrep import ClientResponse, RequestInfo
 
 
 def test_del():
@@ -425,3 +425,43 @@ def test_charset_no_charset():
     response.headers = {'Content-Type': 'application/json'}
 
     assert response.charset is None
+
+
+def test_response_request_info():
+    url = 'http://def-cl-resp.org'
+    headers = {'Content-Type': 'application/json;charset=cp1251'}
+    response = ClientResponse(
+        'get', URL(url),
+        request_info=RequestInfo(
+            url,
+            headers
+        )
+    )
+    assert url == response.request_info.url
+    assert headers == response.request_info.headers
+
+
+def test_response_request_info_empty():
+    url = 'http://def-cl-resp.org'
+    response = ClientResponse(
+        'get', URL(url),
+    )
+    assert response.request_info is None
+
+
+def test_request_info_in_exception():
+    url = 'http://def-cl-resp.org'
+    headers = {'Content-Type': 'application/json;charset=cp1251'}
+    response = ClientResponse(
+        'get',
+        URL(url),
+        request_info=RequestInfo(
+            url,
+            headers
+        )
+    )
+    response.status = 409
+    response.reason = 'CONFLICT'
+    with pytest.raises(aiohttp.ClientResponseError) as cm:
+        response.raise_for_status()
+    assert cm.value.request_info == response.request_info